requirements.txt
langchain-openai
python-dotenv
PyPDF2
pydantic
browser-use
playwright
fastapi
uvicorn